# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Team Health Check API — Pulse Framework
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /check-ins:
    get:
      operationId: getCheckIns
      tags: [check-ins]
      summary: List user check-ins
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckIn"
    post:
      operationId: createCheckIn
      tags: [check-ins]
      summary: Start a new check-in session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckInBody"
      responses:
        "201":
          description: Check-in created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
  /check-ins/{id}:
    get:
      operationId: getCheckIn
      tags: [check-ins]
      summary: Get a single check-in with responses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInDetail"
        "404":
          description: Not found
  /check-ins/{id}/responses:
    post:
      operationId: submitResponses
      tags: [check-ins]
      summary: Submit responses for a check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResponsesBody"
      responses:
        "200":
          description: Responses saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInDetail"
  /check-ins/{id}/complete:
    post:
      operationId: completeCheckIn
      tags: [check-ins]
      summary: Mark a check-in as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
...
  /allowed-emails:
    get:
      operationId: listAllowedEmails
      tags: [allowed-emails]
      summary: List team allowed emails (directors only)
      responses:
        "200":
          description: List of allowed emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllowedEmail"
    post:
      operationId: createAllowedEmail
      tags: [allowed-emails]
      summary: Add an email to the team allowlist (directors only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllowedEmailBody"
      responses:
        "201":
          description: Email added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowedEmail"
  /allowed-emails/{id}:
    delete:
      operationId: deleteAllowedEmail
      tags: [allowed-emails]
      summary: Remove an email from the team allowlist (directors only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CheckIn:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cadence:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, userId, cadence, status, createdAt, completedAt]
    CreateCheckInBody:
      type: object
      properties:
        cadence:
          type: string
          default: weekly
      required: []
    CheckInResponse:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        numericValue:
          type: ["number", "null"]
        textValue:
          type: ["string", "null"]
        emojiValue:
          type: ["string", "null"]
        selectedOptions:
          type: ["array", "null"]
          items:
            type: string
        trafficLight:
          type: ["string", "null"]
          enum: [green, yellow, red, null]
        normalizedScore:
          type: ["number", "null"]
      required: [id, questionId, numericValue, textValue, emojiValue, selectedOptions, trafficLight, normalizedScore]
    CheckInDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cadence:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        responses:
          type: array
          items:
            $ref: "#/components/schemas/CheckInResponse"
      required: [id, userId, cadence, status, createdAt, completedAt, responses]
    SubmitResponsesBody:
      type: object
      properties:
        responses:
          type: array
          items:
            $ref: "#/components/schemas/ResponseInput"
      required: [responses]
    ResponseInput:
      type: object
      properties:
        questionId:
          type: integer
        numericValue:
          type: ["number", "null"]
        textValue:
          type: ["string", "null"]
        emojiValue:
          type: ["string", "null"]
        selectedOptions:
          type: ["array", "null"]
          items:
            type: string
        trafficLight:
          type: ["string", "null"]
          enum: [green, yellow, red, null]
      required: [questionId]
...
    AllowedEmail:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        teamId:
          type: ["integer", "null"]
        invitedByUserId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, email, teamId, invitedByUserId, createdAt]
    CreateAllowedEmailBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]

    # ─── Design Ops: Dashboard ─────────────────────────────────────────
...